🐛 Fix: [物件管理> 新增] 1-6 [物件管理>刪除] ( 管理處 tab 下 ) 特定使用者( User id =1 )才能刪除

+++ app/Http/Controllers/BuildingController.php
@@ -33,7 +33,8 @@ class BuildingController extends Controller
 
         $responseData = new NestedRelationResponser();
         $responseData
-            ->index('buildings',
+            ->index(
+                'buildings',
                 $this->limitRecords(Building::with($request->withNested))
             )
             ->relations($request->withNested);
@@ -49,11 +50,15 @@ class BuildingController extends Controller
     public function create(Request $request)
     {
         $responseData = new FormDataResponser();
-        $data = $responseData
+        $responseData = $responseData
             ->create(Building::class, 'buildings.store')
             ->get();
 
-        return view('buildings.form', $data);
+        if ($request->old()) {
+            $responseData['data'] = array_merge($responseData['data'], $request->old());
+        }
+
+        return view('buildings.form', $responseData);
     }
 
     /**
@@ -67,7 +72,7 @@ class BuildingController extends Controller
         $validatedData = $request->validate([
             'title' => 'required|max:255',
             'building_code' => 'required|max:255',
-            'group' => 'required|max:255',
+            'group' => 'max:255',
             'city' => [
                 'required',
                 Rule::in(array_keys(config('enums.cities')))
@@ -75,47 +80,49 @@ class BuildingController extends Controller
             'district' => [
                 'bail',
                 'required_with:city',
-                Rule::in(config('enums.cities.' . $request->city))
+                Rule::in(config('enums.cities.'.$request->city))
             ],
             'address' => 'required|max:255',
+
+            'is_squatter' => 'boolean',
+            'squatter_status' => 'max:255',
+
             'tax_number' => 'required|max:255',
-            'building_type' => 'required|max:255',
+            'building_type' => 'max:255',
             'floor' => 'required|integer|digits_between:1,11',
-            'legal_usage' => 'required|max:255',
-            'has_elevator' => 'required|boolean',
-            'security_guard' => 'required|max:255',
+            'legal_usage' => 'max:255',
+            'land_use' => 'max:255',
+            'has_elevator' => 'boolean',
+            'security_guard' => 'max:255',
             'management_count' => 'required|max:255',
-            'first_floor_door_opening' => 'required|max:255',
-            'public_area_door_opening' => 'required|max:255',
-            'room_door_opening' => 'required|max:255',
+            'first_floor_door_opening' => 'max:255',
+            'public_area_door_opening' => 'max:255',
+            'room_door_opening' => 'max:255',
             'main_ammeter_location' => 'required|max:255',
             'ammeter_serial_number_1' => 'required|max:255',
-            'shared_electricity' => 'required|max:255',
-            'taiwan_electricity_payment_method' => 'required|max:255',
+            'shared_electricity' => 'max:255',
+            'taiwan_electricity_payment_method' => 'max:255',
             'electricity_payment_method' => [
-                'required',
-                Rule::in(
-                    config('enums.buildings.electricity_payment_method')
-                )
+                Rule::in(config('enums.buildings.electricity_payment_method'))
             ],
-            'private_ammeter_location' => 'required|max:255',
+            'private_ammeter_location' => 'max:255',
             'water_meter_location' => 'required|max:255',
             'water_meter_serial_number' => 'required|max:255',
             'water_payment_method' => 'required|max:255',
             'water_meter_reading_date' => 'required|date',
             'gas_meter_location' => 'required|max:255',
-            'garbage_collection_location' => 'required|max:255',
-            'garbage_collection_time' => 'required|max:255',
-            'management_fee_payment_method' => 'required|max:255',
+            'garbage_collection_location' => 'max:255',
+            'garbage_collection_time' => 'max:255',
+            'management_fee_payment_method' => 'max:255',
             'management_fee_contact' => 'required|max:255',
             'management_fee_contact_phone' => 'required|max:255',
-            'distribution_method' => 'required|max:255',
+            'distribution_method' => 'max:255',
             'administrative_number' => 'required|max:255',
-            'accounting_group' => 'required|max:255',
-            'rental_receipt' => 'required|max:255',
-            'commissioner_id' => 'nullable|exists:users,id',
-            'administrator_id' => 'nullable|exists:users,id',
-            'comment' => 'required|max:255'
+            'accounting_group' => 'max:255',
+            'rental_receipt' => 'max:255',
+            'commissioner_id' => 'exists:users,id',
+            'administrator_id' => 'exists:users,id',
+            'comment' => 'max:255'
         ]);
 
         $newBuilding = BuildingService::create($validatedData);
@@ -146,13 +153,16 @@ class BuildingController extends Controller
      * @param  \App\Building  $building
      * @return \Illuminate\Http\Response
      */
-    public function edit(Building $building)
+    public function edit(Request $request, Building $building)
     {
         $responseData = new FormDataResponser();
-        return view(
-            'buildings.form',
-            $responseData->edit($building, 'buildings.update')->get()
-        );
+        $responseData = $responseData->edit($building, 'buildings.update')->get();
+
+        if ($request->old()) {
+            $responseData['data'] = array_merge($responseData['data'], $request->old());
+        }
+
+        return view('buildings.form', $responseData);
     }
 
     /**
@@ -167,7 +177,7 @@ class BuildingController extends Controller
         $validatedData = $request->validate([
             'title' => 'required|max:255',
             'building_code' => 'required|max:255',
-            'group' => 'required|max:255',
+            'group' => 'max:255',
             'city' => [
                 'required',
                 Rule::in(array_keys(config('enums.cities')))
@@ -175,47 +185,49 @@ class BuildingController extends Controller
             'district' => [
                 'bail',
                 'required_with:city',
-                Rule::in(config('enums.cities.' . $request->city))
+                Rule::in(config('enums.cities.'.$request->city))
             ],
             'address' => 'required|max:255',
+
+            'is_squatter' => 'boolean',
+            'squatter_status' => 'max:255',
+
             'tax_number' => 'required|max:255',
-            'building_type' => 'required|max:255',
+            'building_type' => 'max:255',
             'floor' => 'required|integer|digits_between:1,11',
-            'legal_usage' => 'required|max:255',
-            'has_elevator' => 'required|boolean',
-            'security_guard' => 'required|max:255',
+            'legal_usage' => 'max:255',
+            'land_use' => 'max:255',
+            'has_elevator' => 'boolean',
+            'security_guard' => 'max:255',
             'management_count' => 'required|max:255',
-            'first_floor_door_opening' => 'required|max:255',
-            'public_area_door_opening' => 'required|max:255',
-            'room_door_opening' => 'required|max:255',
+            'first_floor_door_opening' => 'max:255',
+            'public_area_door_opening' => 'max:255',
+            'room_door_opening' => 'max:255',
             'main_ammeter_location' => 'required|max:255',
             'ammeter_serial_number_1' => 'required|max:255',
-            'shared_electricity' => 'required|max:255',
-            'taiwan_electricity_payment_method' => 'required|max:255',
+            'shared_electricity' => 'max:255',
+            'taiwan_electricity_payment_method' => 'max:255',
             'electricity_payment_method' => [
-                'required',
-                Rule::in(
-                    config('enums.buildings.electricity_payment_method')
-                )
+                Rule::in(config('enums.buildings.electricity_payment_method'))
             ],
-            'private_ammeter_location' => 'required|max:255',
+            'private_ammeter_location' => 'max:255',
             'water_meter_location' => 'required|max:255',
             'water_meter_serial_number' => 'required|max:255',
             'water_payment_method' => 'required|max:255',
             'water_meter_reading_date' => 'required|date',
             'gas_meter_location' => 'required|max:255',
-            'garbage_collection_location' => 'required|max:255',
-            'garbage_collection_time' => 'required|max:255',
-            'management_fee_payment_method' => 'required|max:255',
+            'garbage_collection_location' => 'max:255',
+            'garbage_collection_time' => 'max:255',
+            'management_fee_payment_method' => 'max:255',
             'management_fee_contact' => 'required|max:255',
             'management_fee_contact_phone' => 'required|max:255',
-            'distribution_method' => 'required|max:255',
+            'distribution_method' => 'max:255',
             'administrative_number' => 'required|max:255',
-            'accounting_group' => 'required|max:255',
-            'rental_receipt' => 'required|max:255',
-            'commissioner_id' => 'nullable|exists:users,id',
-            'administrator_id' => 'nullable|exists:users,id',
-            'comment' => 'required|max:255'
+            'accounting_group' => 'max:255',
+            'rental_receipt' => 'max:255',
+            'commissioner_id' => 'exists:users,id',
+            'administrator_id' => 'exists:users,id',
+            'comment' => 'max:255'
         ]);
 
         $building->update($validatedData);
@@ -229,13 +241,27 @@ class BuildingController extends Controller
      * @param  \App\Building  $building
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Building $building)
+    public function destroy(Request $request, Building $building)
     {
+        $user = auth()->user();
+
+        $permission = $user->getAllPermissions()->first(function ($permission) {
+            return $permission->name = 'delete building';
+        });
+
+        if (! $permission) {
+            return response()->json([
+                'errors' => ['permission denied']
+            ], 403);
+        }
+
         $building->delete();
+
         return response()->json(true);
     }
 
-    public function electricityPaymentReport(Building $building, int $year, int $month) {
+    public function electricityPaymentReport(Building $building, int $year, int $month)
+    {
         $reportRows = $this->buildElectricityPaymentReportData($building, $year, $month);
 
         return view('buildings.electricity_payment_report', [
@@ -254,9 +280,12 @@ class BuildingController extends Controller
      *
      * @return array
      */
-    private function buildElectricityPaymentReportData(Building $building, int $year, int $month): Collection {
+    private function buildElectricityPaymentReportData(Building $building, int $year, int $month): Collection
+    {
         $rooms = $building->rooms()->get();
 
-        return $rooms->map(function ($room) use ($year, $month) { return $room->buildElectricityPaymentData($year, $month); });
+        return $rooms->map(function ($room) use ($year, $month) {
+            return $room->buildElectricityPaymentData($year, $month);
+        });
     }
 }
